+++ Config.lua
@@ -295,7 +295,7 @@ config:SetScript("OnShow", function(self)
 	fsText:SetText(L["Strata of panel"])
 
 	-- FRAME STRATA COMBOBOX
-	local fsCombobox =  CreateFrame("FRAME", "HidingBarAddonFrameStrata", self.generalPanel, "UIDropDownMenuTemplate")
+	local fsCombobox = CreateFrame("FRAME", "HidingBarAddonFrameStrata", self.generalPanel, "UIDropDownMenuTemplate")
 	fsCombobox:SetPoint("LEFT", fsText, "RIGHT", -12, 0)
 	UIDropDownMenu_SetWidth(fsCombobox, 100)
 
@@ -616,14 +616,14 @@ config:SetScript("OnShow", function(self)
 	-- POSITION BAR PANEL
 	self.positionBarPanel = createTabPanel(settingsTabs, L["Bar position"])
 
-	local function updateTypePosition()
-		self.hidingBar:setBarPosition()
-
+	local function updateFreeMoveChilds()
 		if self.config.freeMove then
 			UIDropDownMenu_EnableDropDown(self.hideToCombobox)
 		else
 			UIDropDownMenu_DisableDropDown(self.hideToCombobox)
 		end
+		self.coordX:SetEnabled(self.config.freeMove)
+		self.coordY:SetEnabled(self.config.freeMove)
 	end
 
 	-- ATTACHED TO THE SIDE
@@ -633,11 +633,11 @@ config:SetScript("OnShow", function(self)
 	self.attachedToSide.Text:SetText(L["Bar attached to the side"])
 	self.attachedToSide:SetScript("OnClick", function(btn)
 		PlaySound(SOUNDKIT.IG_MAINMENU_OPTION_CHECKBOX_ON)
-		self.config.secondPosition = 0
 		self.config.freeMove = nil
 		self.freeMove.check:Hide()
 		btn.check:Show()
-		updateTypePosition()
+		updateFreeMoveChilds()
+		self.hidingBar:setBarPosition(nil, 0)
 	end)
 
 	-- MOVES FREELY
@@ -650,7 +650,7 @@ config:SetScript("OnShow", function(self)
 		self.config.freeMove = true
 		self.attachedToSide.check:Hide()
 		btn.check:Show()
-		updateTypePosition()
+		updateFreeMoveChilds()
 	end)
 
 	-- HIDE TO
@@ -694,8 +694,75 @@ config:SetScript("OnShow", function(self)
 	end)
 	UIDropDownMenu_SetSelectedValue(self.hideToCombobox, self.config.anchor)
 
-	-- ENABLE OR DISABLE HIDE TO
-	updateTypePosition()
+	-- COORD X
+	self.coordXText = self.positionBarPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
+	self.coordXText:SetPoint("LEFT", self.hideToCombobox, "RIGHT", -5, 2)
+	self.coordXText:SetText("X")
+
+	self.coordX = CreateFrame("EditBox", nil, self.positionBarPanel, "HidingBarAddonCoordTextBox")
+	self.coordX:SetPoint("LEFT", self.coordXText, "RIGHT", 1, 0)
+	self.coordX:SetScript("OnTextChanged", function(editBox, userInput)
+		if userInput then
+			editBox:SetNumber(editBox:GetText():match("%-?%d*"))
+		end
+	end)
+	self.coordX.setX = function(editBox, x)
+		if self.config.anchor == "left" or self.config.anchor == "right" then
+			self.hidingBar:setBarPosition(nil, x)
+		else
+			self.hidingBar:setBarPosition(x)
+		end
+	end
+	self.coordX:SetScript("OnEnterPressed", function(editBox)
+		editBox:setX(tonumber(editBox:GetText():match("%-?%d*")) or 0)
+		editBox:ClearFocus()
+	end)
+	self.coordX:SetScript("OnEditFocusLost", function(editBox)
+		self:updateCoords()
+		editBox:HighlightText(0, 0)
+	end)
+	self.coordX:SetScript("OnMouseWheel", function(editBox, delta)
+		local int = tonumber(editBox:GetText():match("%-?%d*")) or 0
+		editBox:setX(int + (delta > 0 and 1 or -1))
+	end)
+
+	-- COORD Y
+	self.coordYText = self.positionBarPanel:CreateFontString(nil, "ARTWORK", "GameFontHighlight")
+	self.coordYText:SetPoint("LEFT", self.coordX, "RIGHT", 5, 0)
+	self.coordYText:SetText("Y")
+
+	self.coordY = CreateFrame("EditBox", nil, self.positionBarPanel, "HidingBarAddonCoordTextBox")
+	self.coordY:SetPoint("LEFT", self.coordYText, "RIGHT", 1, 0)
+	self.coordY:SetScript("OnTextChanged", function(editBox, userInput)
+		if userInput then
+			editBox:SetNumber(editBox:GetText():match("%-?%d*"))
+		end
+	end)
+	self.coordY.setY = function(editBox, y)
+		if self.config.anchor == "left" or self.config.anchor == "right" then
+			self.hidingBar:setBarPosition(y)
+		else
+			self.hidingBar:setBarPosition(nil, y)
+		end
+	end
+	self.coordY:SetScript("OnEnterPressed", function(editBox)
+		editBox:setY(tonumber(editBox:GetText():match("%-?%d*")) or 0)
+		editBox:ClearFocus()
+	end)
+	self.coordY:SetScript("OnEditFocusLost", function(editBox)
+		self:updateCoords()
+		editBox:HighlightText(0, 0)
+	end)
+	self.coordY:SetScript("OnMouseWheel", function(editBox, delta)
+		local int = tonumber(editBox:GetText():match("%-?%d*")) or 0
+		editBox:setY(int + (delta > 0 and 1 or -1))
+	end)
+
+	-- SET COORDS
+	self:updateCoords()
+
+	-- ENABLE OR DISABLE "MOVES FREELY" OPTIONS
+	updateFreeMoveChilds()
 
 	-- BUTTONS TAB PANEL
 	self.buttonsTabPanel = createTabPanel(buttonTabs, L["Buttons"])
@@ -812,6 +879,20 @@ config:SetScript("OnShow", function(self)
 		self:applyLayout()
 	end)
 end)
+
+
+function config:updateCoords()
+	if not self.coordX or not self.coordY then return end
+
+	local scale = UIParent:GetScale()
+	local x, y = self.config.position / scale, self.config.secondPosition / scale
+	if self.config.anchor == "left" or self.config.anchor == "right" then
+		x, y = y, x
+	end
+	
+	self.coordX:SetNumber(math.floor(x + .5))
+	self.coordY:SetNumber(math.floor(y + .5))
+end
 
 
 function config:addIgnoreName(name)
@@ -1088,7 +1169,7 @@ function config:applyLayout(delay)
 
 	local rows = math.ceil((#self.mbuttons + self.orderMBtnDelta) / self.size)
 	local width = self.size * self.config.buttonSize + offsetX * 2
-	local height = rows * self.config.buttonSize +  offsetY * 2
+	local height = rows * self.config.buttonSize + offsetY * 2
 	self.buttonPanel:SetSize(width, height)
 end